Revert "just getting everything in until I move -- might delete later"
This reverts commit 37efacd77379c229d1224a02e260c3632ab9884a.

Note for future?: The above commit happened because I was moving countries
and assumed i'd be getting a new computer, so I pushed everything I was
working on, with intention of reverting later. New computer didn't happen so
the project was abandoned. Revert is just so that last commit is a
working commit.